back/pyproject.toml
[project]
name = "back"
version = "0.1.0"
description = ""
authors = [
    {name = "kensho.takizawa", email = "kensho.takizawa@minedia.com"}
]
readme = "README.md"
requires-python = "^3.9"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
fastapi = ">=0.115.8,<0.116.0"
uvicorn = ">=0.34.0,<0.35.0"
openai = ">=1.60.2,<2.0.0"
python-dotenv = ">=1.0.1,<2.0.0"
starlette = ">=0.40.0,<0.46.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn main:app --reload"